Interesting strategy for Parser to be less annoying...
Now instead of using references in PoolS, I'm just using raw
pointers. This is technically unsafe, but it does seem to be fine?

I'm just not sure how to otherwise wrangle the lifetimes so that
Parser will know that it lives long enough so that the Tokens won't
cause a problem, without borrowing the Parser for so long that
it's impossible to repeatedly advance() it.