Start working on "Bindings"
The bindings idea is going to be needed to implement type inference
as well as top-level order independence.

Basically, when we see an identifier, we MIGHT not have seen it yet
(because unlike C, everything may be rearranged in any order...)

So, we just have to keep track that we have an resolved binding.
Once it is bound, there is a specific globally unique object that
is associated with a specific variable/function access.

One other thing: Because, within functions, variables do need
to be kept track of in linear order, what we can do is simply
keep track of bindings as we find them inside a function and
then completely forget about them once we stop building that
part of the syntax tree....

This means that, at the time the Parser finishes, if there
are any unbound variables or functions, they MUST be "global".