Get numeric type checking working rather nicely
This is actually pretty cool. Basically, if a number constant doesn't
have a specific type, the type system will try to automatically
give it one.

This does have one problem, which is it won't work as well with
e.g. 10.5, something with a decimal point... but we can force all
those types to be floats for now, and decide whether to have
separate doubles later (and then add an UnspecificFloat type
with similar machinery to UnspecificNumeric).