Make typecheck change AST from Untyped to Typed
This change means that, after the typecheck stage, the AST is
guaranteed to have valid bindings for every single node. This is
an improvement, as it makes writing the codegen easier.

However, the one problem it introduces is that some values are
basically being un-boxed and then later re-boxed. This could
introduce a bunch of technically unnecessary allocations.
There may be a way to tell Rust to re-use the boxes?

That is, if the type is smaller or the same size, it seems
like it should be able to turn a Box<A> into a Box<B> using
a function that turns A into B.